requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==4.2.6
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-cors-headers==4.3.0
django-crispy-forms==2.1
django-filter==23.3
django-js-asset==2.1.0
django-phonenumber-field==7.2.0
django-resized==1.0.2
djangorestframework==3.14.0
drf-writable-nested==0.7.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
phonenumbers==8.13.22
Pillow==10.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
